update code to work with Coppelia Sim 4.6.0. Not working yet. VREP 3.6.2 untested

+++ EAFramework/include/ARE/Environment.h
@@ -16,8 +16,8 @@
 namespace are {
 
 struct waypoint{
-    float position[3];
-    float orientation[3];
+    double position[3];
+    double orientation[3];
 
     std::string to_string() const{
         std::stringstream sstr;
